Refactor CLI and setup fetch-params command
