+++ forest/src/cli/mod.rs
@@ -16,22 +16,32 @@ use std::sync::Arc;
 use structopt::StructOpt;
 use utils::{read_file_to_string, read_toml};
 
-#[derive(Debug, StructOpt)]
+/// CLI structure generated when interacting with Forest binary
+#[derive(StructOpt)]
 #[structopt(
-    name = "Forest",
+    name = "forest",
     version = "0.0.1",
-    about = "Filecoin implementation in Rust",
+    about = "Filecoin implementation in Rust. This command will start the daemon process",
     author = "ChainSafe Systems <info@chainsafe.io>"
 )]
 pub struct CLI {
+    #[structopt(flatten)]
+    pub daemon_opts: DaemonOpts,
+    #[structopt(subcommand)]
+    pub cmd: Option<Subcommand>,
+}
+
+/// Daemon process command line options.
+#[derive(StructOpt, Debug)]
+pub struct DaemonOpts {
     #[structopt(short, long, help = "A toml file containing relevant configurations.")]
     pub config: Option<String>,
     #[structopt(short, long, help = "The genesis CAR file")]
     pub genesis: Option<String>,
 }
 
-impl CLI {
-    pub fn get_config(&self) -> Result<Config, io::Error> {
+impl DaemonOpts {
+    pub fn to_config(&self) -> Result<Config, io::Error> {
         let mut cfg: Config = match &self.config {
             Some(config_file) => {
                 // Read from config file
@@ -48,6 +58,19 @@ impl CLI {
 
         Ok(cfg)
     }
+}
+
+/// Forest binary subcommands available.
+#[derive(StructOpt)]
+pub enum Subcommand {
+    #[structopt(
+        name = "fetch-params",
+        about = "Download parameters for generating and verifying proofs for given size"
+    )]
+    FetchParams {
+        #[structopt(help = "Size in bytes")]
+        params_size: i64,
+    },
 }
 
 /// Blocks current thread until ctrl-c is received
