+++ forest/src/subcommand.rs
@@ -1,0 +1,9 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use super::cli::Subcommand;
+
+/// Process CLI subcommand
+pub(super) fn process(_command: Subcommand) {
+    todo!()
+}
